+++ src/main/java/com/prgrms/boardjpa/application/user/exception/AuthenticationFailException.java
@@ -1,0 +1,7 @@
+package com.prgrms.boardjpa.application.user.exception;
+
+public class AuthenticationFailException extends RuntimeException{
+	public AuthenticationFailException() {
+
+	}
+}
