+++ src/main/java/com/prgrms/boardjpa/application/user/exception/AuthorizationFailException.java
@@ -1,0 +1,7 @@
+package com.prgrms.boardjpa.application.user.exception;
+
+public class AuthorizationFailException extends RuntimeException {
+	public AuthorizationFailException() {
+		//TODO
+	}
+}
