+++ src/main/java/com/prgrms/boardjpa/application/user/model/Hobby.java
@@ -1,0 +1,8 @@
+package com.prgrms.boardjpa.application.user.model;
+
+public enum Hobby {
+	MOVIE,
+	PICTURE,
+	WALKING,
+	NONE
+}
