--- src/main/java/com/prgrms/boardjpa/commons/api/ErrorResponse.java
@@ -1,13 +1,0 @@
-package com.prgrms.boardjpa.commons.api;
-
-public class ErrorResponse {
-	private final String errorMessage;
-
-	public ErrorResponse(Throwable throwable) {
-		this.errorMessage = throwable.getMessage();
-	}
-
-	public String getErrorMessage() {
-		return errorMessage;
-	}
-}
