--- src/main/java/com/prgrms/boardjpa/commons/api/SuccessResponse.java
@@ -1,17 +1,0 @@
-package com.prgrms.boardjpa.commons.api;
-
-public class SuccessResponse<T> {
-	private final T data;
-
-	private SuccessResponse(T data) {
-		this.data = data;
-	}
-
-	public static <T> SuccessResponse<T> of(T data) {
-		return new SuccessResponse<>(data);
-	}
-
-	public T getData() {
-		return data;
-	}
-}
