--- src/main/java/com/prgrms/boardjpa/commons/exception/NotExistException.java
@@ -1,7 +1,0 @@
-package com.prgrms.boardjpa.commons.exception;
-
-public class NotExistException extends RuntimeException {
-	public NotExistException() {
-		super("존재하지 않습니다");
-	}
-}
