+++ src/main/java/com/prgrms/boardjpa/core/commons/exception/CreationFailException.java
@@ -1,0 +1,7 @@
+package com.prgrms.boardjpa.core.commons.exception;
+
+public class CreationFailException extends RuntimeException {
+	public CreationFailException(Class<?> clazz, Throwable cause) {
+		super(clazz.getSimpleName() + "생성에 실패하였습니다", cause);
+	}
+}
