+++ src/main/java/com/prgrms/boardjpa/core/config/JpaAuditingConfig.java
@@ -1,0 +1,9 @@
+package com.prgrms.boardjpa.core.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@Configuration
+@EnableJpaAuditing
+public class JpaAuditingConfig {
+}
