--- src/main/java/com/prgrms/boardjpa/user/domain/Hobby.java
@@ -1,8 +1,0 @@
-package com.prgrms.boardjpa.user.domain;
-
-public enum Hobby {
-	MOVIE,
-	PICTURE,
-	WALKING,
-	NONE
-}
