--- src/main/java/com/prgrms/boardjpa/user/exception/AuthenticationFailException.java
@@ -1,7 +1,0 @@
-package com.prgrms.boardjpa.user.exception;
-
-public class AuthenticationFailException extends RuntimeException{
-	public AuthenticationFailException() {
-
-	}
-}
