--- src/main/java/com/prgrms/boardjpa/user/exception/AuthorizationFailException.java
@@ -1,7 +1,0 @@
-package com.prgrms.boardjpa.user.exception;
-
-public class AuthorizationFailException extends RuntimeException {
-	public AuthorizationFailException() {
-		//TODO
-	}
-}
